nokkrar lagfæringar

+++ src/main/java/is/hi/hbv501g/hbv1/Services/Implementations/TypingServiceImplementation.java
@@ -115,6 +115,9 @@ public class TypingServiceImplementation implements TypingService {
     @Override
     public void deleteQuote(Quote quote) {
         Quote quote1 = getQuoteById(quote.getId());
+        if (quote1.isAccepted()){
+            return;
+        }
         quoteRepository.delete(quote1);
     }
